Moved variantGroups into genome browser's constants directory
