+++ src/content/app/genome-browser/components/drawer/components/variant-colour/VariantColour.tsx
@@ -18,7 +18,7 @@ import React from 'react';
 import classNames from 'classnames';
 
 // QUESTION: should variantGroups be moved to a different directory? E.g. to genome-browser/constants?
-import variantGroups from 'src/content/app/genome-browser/components/drawer/drawer-views/variant-group-legend/variantGroups';
+import variantGroups from 'src/content/app/genome-browser/constants/variantGroups';
 
 import styles from './VariantColour.scss';
