+++ src/content/app/genome-browser/constants/variantGroups.ts
